+++ src/main/java/org/sourcelab/storm/spout/redis/client/ClientType.java
@@ -1,0 +1,9 @@
+package org.sourcelab.storm.spout.redis.client;
+
+/**
+ * Defines allowed implementations.
+ */
+public enum ClientType {
+    LETTUCE,
+    JEDIS;
+}