+++ src/main/java/org/sourcelab/storm/spout/redis/client/lettuce/LettuceClient.java
@@ -1,4 +1,4 @@
-package org.sourcelab.storm.spout.redis.client;
+package org.sourcelab.storm.spout.redis.client.lettuce;
 
 import io.lettuce.core.Consumer;
 import io.lettuce.core.RedisBusyException;
@@ -12,6 +12,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.sourcelab.storm.spout.redis.Message;
 import org.sourcelab.storm.spout.redis.RedisStreamSpoutConfig;
+import org.sourcelab.storm.spout.redis.client.Client;
 
 import java.util.List;
 import java.util.Objects;
@@ -44,7 +45,13 @@ public class LettuceClient implements Client {
      */
     private final XReadArgs xreadArgs;
     private final Consumer<String> consumerFrom;
-    private final XReadArgs.StreamOffset<String> lastConsumed;
+
+    /**
+     * State for consuming first from consumer's personal pending list,
+     * then switching to reading from consumer group messages.
+     */
+    private boolean hasFinishedPpl = false;
+    private XReadArgs.StreamOffset<String> lastConsumed;
 
     /**
      * Constructor.
@@ -55,11 +62,8 @@ public class LettuceClient implements Client {
         this(
             config,
             instanceId,
-            // Determine which adapter to use based on what type of redis instance we are
-            // communicating with.
-            config.isConnectingToCluster()
-                ? new LettuceClusterAdapter(RedisClusterClient.create(config.getConnectString()))
-                : new LettuceRedisAdapter(RedisClient.create(config.getConnectString()))
+            // Determine which adapter to use based on what type of redis instance we are communicating with.
+            createAdapter(config)
         );
     }
 
@@ -85,9 +89,6 @@ public class LettuceClient implements Client {
 
         // Create re-usable ConsumerFrom instance.
         consumerFrom = Consumer.from(config.getGroupName(), consumerId);
-
-        // Create re-usable lastConsumed instance.
-        lastConsumed = XReadArgs.StreamOffset.lastConsumed(config.getStreamKey());
     }
 
     @Override
@@ -97,6 +98,10 @@ public class LettuceClient implements Client {
         }
 
         adapter.connect();
+
+        // Create re-usable lastConsumed instance, default to consuming from PPL list
+        lastConsumed = XReadArgs.StreamOffset.from(config.getStreamKey(), "0-0");
+        hasFinishedPpl = false;
 
         try {
             // Attempt to create consumer group
@@ -128,17 +133,35 @@ public class LettuceClient implements Client {
     @Override
     public List<Message> nextMessages() {
         // Get next batch of messages.
-        final List<StreamMessage<String, String>> messages = adapter.getSyncCommands().xreadgroup(
+        final List<StreamMessage<String, String>> entries = adapter.getSyncCommands().xreadgroup(
             consumerFrom,
             xreadArgs,
             lastConsumed
         );
 
         // Loop over each message
-        return messages.stream()
+        final List<Message> messages = entries.stream()
             // Map into Message Object
             .map((streamMsg) -> new Message(streamMsg.getId(), streamMsg.getBody()))
             .collect(Collectors.toList());
+
+        if (!hasFinishedPpl) {
+            if (messages.isEmpty()) {
+                logger.info("Personal Pending List appears empty, switching to consuming from new messages.");
+
+                hasFinishedPpl = true;
+                lastConsumed = XReadArgs.StreamOffset.lastConsumed(config.getStreamKey());
+
+                // Re-attempt consuming
+                return nextMessages();
+            } else {
+                // Advance last index consumed from PPL so we don't continue to replay old messages.
+                final String lastId = messages.get(messages.size() - 1).getId();
+                lastConsumed = XReadArgs.StreamOffset.from(config.getStreamKey(), lastId);
+            }
+        }
+
+        return messages;
     }
 
     @Override
@@ -154,5 +177,20 @@ public class LettuceClient implements Client {
     @Override
     public void disconnect() {
         adapter.shutdown();
+    }
+
+    /**
+     * Factory method for creating the appropriate adapter based on configuration.
+     * @param config Spout configuration.
+     * @return Appropriate Adapter.
+     */
+    private static LettuceAdapter createAdapter(final RedisStreamSpoutConfig config) {
+        if (config.isConnectingToCluster()) {
+            logger.info("Connecting to RedisCluster at {}", config.getConnectStringMasked());
+            return new LettuceClusterAdapter(RedisClusterClient.create(config.getConnectString()));
+        } else {
+            logger.info("Connecting to Redis server at {}", config.getConnectStringMasked());
+            return new LettuceRedisAdapter(RedisClient.create(config.getConnectString()));
+        }
     }
 }